Changed most value types to structs
Many classes had only a value character and so it is better to declare them as struct. That was done with some classes with the exception of ReferenceStep and FlowIdentifier because they would break some code in the XMLParser module. Also the equality and hash methdos for those classes were overriden.